Windows compilation fixes + bump version.

+++ Source/ARX/OCVT/PlanarTracker.cpp
@@ -478,7 +478,7 @@ public:
         if(fs.isOpened())
         {
             try {
-                int totalTrackables = _trackables.size();
+                int totalTrackables = (int)_trackables.size();
                 fs << "totalTrackables" << totalTrackables;
                 fs << "featureType" << _selectedFeatureDetectorType;
                 for(int i=0;i<_trackables.size(); i++) {
